// core/packages/web-client/src/app/environments/store/types.ts
import { EnvironmentJSON } from "ditto-shared";

export const enum EnvironmentActionTypes {
  ENVIRONMENT_CREATION_STARTED = "ENVIRONMENT_CREATION_STARTED",
  ENVIRONMENT_CREATION_SUCCEEDED = "ENVIRONMENT_CREATION_SUCCEEDED",
  ENVIRONMENT_CREATION_FAILED = "ENVIRONMENT_CREATION_FAILED",
  RECIEVE_ENVIRONMENTS = "RECIEVE_ENVIRONMENTS",
  ENVIRONMENT_SELECTED = "ENVIRONMENT_SELECTED",
  ENVIRONMENT_UPDATE_STARTED = "ENVIRONMENT_UPDATE_STARTED",
  ENVIRONMENT_UPDATE_SUCCEEDED = "ENVIRONMENT_UPDATE_SUCCEEDED",
  ENVIRONMENT_UPDATE_FAILED = "ENVIRONMENT_UPDATE_FAILED",
  ENVIRONMENT_DELETION_STARTED = "ENVIRONMENT_DELETE_STARTED",
  ENVIRONMENT_DELETION_SUCCEEDED = "ENVIRONMENT_DELETE_SUCCEEDED",
  ENVIRONMENT_DELETION_FAILED = "ENVIRONMENT_DELETION_FAILED"
}



export interface EnvironmentsState{

    readonly all : EnvironmentJSON[];
    readonly selected : EnvironmentJSON | undefined;
}